Use KPluginMetaData for json-KPluginInfo handling
This does away with converting to and from QVariantMaps, -Hashes and
Lists, and uses the new-in-5.3 API in KPluginInfo to convert to and from
KPluginMetaData. This means we bother less with the internal structure,
and use the API better.

As a side-effect, this speeds up the whole process a bit.