Fix PluginTrader
We're using the old codepath, to parse directories, basically backing
out the move to KPluginLoader for now, as this seems to brittle.

Also fix the filter() lambda to take empty servicetypes into account and
return true accordingly, this saves doing it in the callers and fixes a
bug.

Now all tests pass, with and without index files.